+++ src/main/java/com/trendyTracker/Appservice/domain/AppInfo.java
@@ -1,0 +1,31 @@
+package com.trendyTracker.Appservice.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+@Entity
+@Table(name = "app_info", schema = "public")
+public class AppInfo {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    private String version;
+
+    @Column(columnDefinition = "TEXT")
+    private String terms;
+
+    // 연관관계 메서드
+    public void addAppInfo(String version, String terms) {
+        this.version = version;
+        this.terms = terms;
+    }
+
+}
