+++ src/main/java/com/trendyTracker/Appservice/repository/AlarmRepository.java
@@ -1,0 +1,9 @@
+package com.trendyTracker.Appservice.repository;
+
+import com.trendyTracker.Appservice.domain.User;
+
+public interface AlarmRepository {
+    void createUser(User user);
+
+    void registTemplate(User user ,String title, String content);
+}
