+++ src/main/java/com/trendyTracker/api/AppInfo/AppInfoController.java
@@ -4,9 +4,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.trendyTracker.Appservice.domain.AppInfo;
+import com.trendyTracker.Appservice.service.AppInfoService;
 import com.trendyTracker.common.config.Loggable;
 import com.trendyTracker.common.response.Response;
-import com.trendyTracker.service.AppInfoService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
