--- src/main/java/com/trendyTracker/repository/AlarmRepository.java
@@ -1,9 +1,0 @@
-package com.trendyTracker.repository;
-
-import com.trendyTracker.domain.AppService.User;
-
-public interface AlarmRepository {
-    void createUser(User user);
-
-    void registTemplate(User user ,String title, String content);
-}
