--- src/main/java/com/trendyTracker/repository/AppInfoRepository.java
@@ -1,11 +1,0 @@
-package com.trendyTracker.repository;
-
-import java.util.Optional;
-
-import com.trendyTracker.domain.AppService.AppInfo;
-
-public interface AppInfoRepository {
-    Optional<AppInfo> getAppInfo();
-
-    void saveAppInfo(String version, String terms);
-}
