+++ src/main/java/com/trendyTracker/util/TechUtils.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import com.trendyTracker.domain.Job.Tech;
+import com.trendyTracker.Job.domain.Tech;
 
 public class TechUtils {
