+++ src/main/java/com/trendyTracker/util/UrlReader.java
@@ -21,7 +21,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.trendyTracker.domain.Job.Tech;
+import com.trendyTracker.Job.domain.Tech;
 
 public class UrlReader {
     static Logger logger = LoggerFactory.getLogger(UrlReader.class);
